+++ src/components/BaseColumn.php
@@ -467,7 +467,6 @@ abstract class BaseColumn extends BaseObject
      */
     protected function renderDragColumn($name, $value, $options)
     {
-        // return Html::tag('span', null, ['class' => 'glyphicon glyphicon-menu-hamburger drag-handle']);
         /**
          * Class was passed into options by TableRenderer->renderCellContent(),
          * we can extract it here
@@ -478,7 +477,7 @@ abstract class BaseColumn extends BaseObject
         }
         $dragClass = implode(" ", [$class, 'drag-handle']);
 
-        return Html::tag('span', null, ['class'=>$dragClass]);
+        return Html::tag('span', null, ['class' => $dragClass]);
     }
 
     /**