+++ src/renderers/BaseRenderer.php
@@ -166,7 +166,7 @@ abstract class BaseRenderer extends BaseObject implements RendererInterface
      */
     public $layoutConfig = [];
 
-    public $fontMap;
+    public $iconMap;
 
     /**
      * @inheritdoc
@@ -250,7 +250,7 @@ abstract class BaseRenderer extends BaseObject implements RendererInterface
         }
 
         if (!array_key_exists('label', $this->removeButtonOptions)) {
-            $this->removeButtonOptions['label'] = Html::tag('i', null, ['class' => $this->getFontClass('remove')]);
+            $this->removeButtonOptions['label'] = Html::tag('i', null, ['class' => $this->getIconClass('remove')]);
         }
 
         if (!array_key_exists('class', $this->addButtonOptions)) {
@@ -258,7 +258,7 @@ abstract class BaseRenderer extends BaseObject implements RendererInterface
         }
 
         if (!array_key_exists('label', $this->addButtonOptions)) {
-            $this->addButtonOptions['label'] = Html::tag('i', null, ['class' => $this->getFontClass('add')]);
+            $this->addButtonOptions['label'] = Html::tag('i', null, ['class' => $this->getIconClass('add')]);
         }
 
         if (!array_key_exists('class', $this->cloneButtonOptions)) {
@@ -266,7 +266,7 @@ abstract class BaseRenderer extends BaseObject implements RendererInterface
         }
 
         if (!array_key_exists('label', $this->cloneButtonOptions)) {
-            $this->cloneButtonOptions['label'] = Html::tag('i', null, ['class' => $this->getFontClass('clone')]);
+            $this->cloneButtonOptions['label'] = Html::tag('i', null, ['class' => $this->getIconClass('clone')]);
         }
     }
 
@@ -496,9 +496,13 @@ abstract class BaseRenderer extends BaseObject implements RendererInterface
      * @param $action - the control parameter, used as key into allowed types
      * @return string - the relevant icon class
      */
-    protected function getFontClass($action) {
+    protected function getIconClass($action) {
         if (in_array($action, ['add', 'remove', 'clone', 'drag-handle'])) {
-            return $this->fontMap[$action];
+            return $this->iconMap[$action];
+        }
+
+        if (YII_DEBUG) {
+            throw new InvalidConfigException('Out of bounds, "' . $action . '" not found in your iconMap');
         }
         return '';
     }