+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Yii2 multiple input change log
 
 2.15.1 (for review)
 =======================
-- #228 added `fontMap` and `fontSource`property for MultipleInput and TabularInput
+- #228 added `iconMap` and `iconSource`property for MultipleInput and TabularInput
 - #228 changed the following methods to support icon class:
     BaseColumn->renderDragColumn(), TableRenderer->renderCellContent(), BaseRenderer->prepareButtons()